Use common CRC checks across both loaders.
